pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "live-research-copilot"
version = "0.1.0"
description = ""
authors = [
  {name = "Farrukh-Olimov", email = "olimov.farrukh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "alembic (>=1.17.2,<2.0.0)",
  "asyncpg (>=0.31.0,<0.32.0)",
  "greenlet (>=3.3.1,<4.0.0)",
  "psycopg2-binary (>=2.9.11,<3.0.0)",
  "python-dotenv (>=1.2.1,<2.0.0)",
  "qdrant-client (>=1.16.1,<2.0.0)",
  "statsd (>=4.0.1,<5.0.0)"
]

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | __pycache__
  | build
  | dist
  | .gitignore
  | .postgres_data 
)/
'''

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = ["pragma: no cover"] # Ignore specific lines
fail_under = 80

[tool.coverage.run]
branch = true
source = ["./common/datasources/", "./common/services/"]
omit = [
  "venv/*",
  ".venv/*",
  ".upload/*",
  "*/__init__.py",
  "*/__main__.py",
  "*/__pycache__/*",
  "*/alembic/*",
  "*/base*.py"
]

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
torch = {source = "pytorch-gpu"}

[tool.poetry.group.ray.dependencies]
apache-airflow = ">=3.1.7,<4.0.0"
poetry-plugin-sort = ">=0.3.0,<0.4.0"
ray = ">=2.52.0,<3.0.0"
scipy = ">=1.16.3,<2.0.0"
sentencepiece = ">=0.2.1,<0.3.0"
torch = "==2.7"
torchaudio = "==2.7"
torchvision = "==0.22"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
black = "^25.1.0"
pytest = "^8.3.5"
pytest-asyncio = "^1.3.0"
pytest-cov = "^6.2.1"
pytest-progress = "^1.4.0"
pytest-xdist = "^3.8.0"
ruff = "^0.14.8"
toml-sort = "^0.24.2"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu128"
priority = "explicit"

[tool.pytest.ini_options]
asyncio_mode = "auto"
# Logging configure
log_cli = false
log_cli_level = "CRITICAL"
log_cli_format = "%(pathname)s:%(lineno)d %(levelname)s %(funcName)s: %(message)s"
log_level = "CRITICAL"
# Basic execition options
addopts = [
  # "-n",   # concurrent tests
  # "auto", # concurrent tests
  # "--dist=load", # loadscope: all test on module or class in same worker, load: functions to workers, loadfile: one test file per worker
  # "-s", # enable output to console
  "-v" # verbose
]

[tool.ruff]
line-length = 88
exclude = [
  ".git",
  "__pycache__",
  ".venv",
  "venv",
  ".postgres_data",
  "pyproject.toml"
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "D"]
# Equivalent of your flake8 ignore
ignore = [
  "E203", # whitespace before ':'
  "D100", # missing docstring in public module
  "D104", # missing docstring in public package
  "D101" # missing docstring in public class
]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.toml-sort]
indent = 4
check = false
format = true
multi = true
line-length = 88
